rename sidebarlayout to applayout
