+++ src/components/AppLayout/Header/components/WalletIcon/icons/icon-authereum.png
