+++ src/components/AppLayout/Header/components/WalletIcon/icons/index.ts
