+++ src/components/AppLayout/Header/components/WalletIcon/index.tsx
